Layout structure updated
