+++ src/routes/(inner)/docs/installation/+page.svelte
@@ -1,0 +1,23 @@
+<script lang="ts">
+	import CodeBlock from '$docs/components/CodeBlock/CodeBlock.svelte';
+
+	// ...
+</script>
+
+<div class="space-y-10">
+	<!-- Header -->
+	<header class="card card-gradient space-y-8">
+		<h1 class="h1"><span>Getting Started</span></h1>
+		<p>
+			Lorem ipsum dolor sit amet consectetur adipisicing elit. Incidunt corporis aut voluptates
+			dolore dignissimos non asperiores officia debitis totam fugiat, ipsam odit repudiandae
+			officiis atque eaque, quod itaque error delectus.
+		</p>
+	</header>
+	<!-- Installation -->
+	<section class="space-y-8">
+		<h2 class="h2">Installation</h2>
+		<p>Call the Hook inside a component.</p>
+		<CodeBlock lang="ts" code={`(snippet)`} />
+	</section>
+</div>
