+++ src/routes/+page.svelte
@@ -1,11 +1,18 @@
 <script lang="ts">
-	import IconArrowRight from 'lucide-svelte/icons/arrow-right';
+	// Components
+	import PageHeader from '$docs/components/PageHeader/PageHeader.svelte';
 	import Logo from '$docs/components/Logo/Logo.svelte';
+	import PageFooter from '$docs/components/PageFooter/PageFooter.svelte';
+	// Icons
+	import IconArrowRight from 'lucide-svelte/icons/arrow-right';
 </script>
 
-<div class="space-y-10">
-	<header
-		class="h-[75vh] flex flex-col justify-center items-center space-y-8 text-center text-balance"
+<div class="h-screen space-y-10">
+	<!-- Header -->
+	<PageHeader />
+	<!-- Content -->
+	<main
+		class="h-[calc(100%-250px)] flex flex-col justify-center items-center space-y-8 text-center text-balance"
 	>
 		<Logo />
 		<h1 class="h1 text-black dark:text-white">Floating UI Svelte</h1>
@@ -17,7 +24,7 @@
 			>
 		</p>
 		<div class="flex gap-4">
-			<a href="/tooltips" class="btn-cta">
+			<a href="/docs/installation" class="btn-cta">
 				<span>Get Started</span>
 				<IconArrowRight />
 			</a>
@@ -29,5 +36,7 @@
 				GitHub
 			</a>
 		</div>
-	</header>
+	</main>
+	<!-- Footer -->
+	<PageFooter />
 </div>
